Fix Muon optimizer using fp16 instead of bf16 on Ampere+ GPUs

+++ train_gpt.py
@@ -100,7 +100,7 @@ def zeropower_via_newtonschulz5(G: Tensor, steps: int = 10, eps: float = 1e-7) -
     # Orthogonalize a 2D update matrix with a fast Newton-Schulz iteration.
     # Muon uses this to normalize matrix-shaped gradients before applying them.
     a, b, c = (3.4445, -4.7750, 2.0315)
-    X = G.half()
+    X = G.to(_amp_dtype)
     X /= X.norm() + eps
     transposed = G.size(0) > G.size(1)
     if transposed:
@@ -140,7 +140,7 @@ class Muon(torch.optim.Optimizer):
             nesterov = group["nesterov"]
 
             total_params = sum(int(p.numel()) for p in params)
-            updates_flat = torch.zeros(total_params, device=params[0].device, dtype=torch.float16)
+            updates_flat = torch.zeros(total_params, device=params[0].device, dtype=_amp_dtype)
 
             curr = 0
             for i, p in enumerate(params):
